+++ packages/angular/testing/src/public-api.ts
@@ -2,9 +2,11 @@ export {
   TisAccordionHarness,
   TisButtonHarness,
   TisCheckboxHarness,
+  TisInputHarness,
   TisModalHarness,
   TisPopoverHarness,
   TisRadioGroupHarness,
   TisSelectHarness,
+  TisTextareaHarness,
   TisToggleHarness,
 } from "./harnesses";